requirements.txt
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-community>=0.0.10
reportlab>=4.0.0
langchain-core>=0.1.0
